+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 * Feature: `EnvState` got more methods to control the envelope (such as triggering/retriggering).
 * Feature: `env_target_stage_lin_time_adj` added, for a retriggerable envelope.
+* Documentation: I've added every piece of foreign code and it's license to the README and lib.rs
+documentation.
 
 0.5.4 (2022-08-28)
 ==================