+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Feature: `EnvState` got more methods to control the envelope (such as triggering/retriggering).
 * Feature: `env_target_stage_lin_time_adj` added, for a retriggerable envelope.
 * Feature: Added an attack decay envelope implementation with `EnvRetrigAD`.
+* Feature: Added `coef2gain_db` and `gain_db2coeff` for converting decibel gain values to coefficients.
 * Documentation: I've added every piece of foreign code and it's license to the README and lib.rs
 documentation.